load setting correct

+++ src/com/hotteststudio/model/UserSettings.java
@@ -8,7 +8,7 @@ public class UserSettings {
 	public int flipper; // 0 Instant 1 Scroller 2 Slider
 	public int brightness; // 10 to 100
 	public int fontType; // list of 3 font
-	public float lineSpacing;
+	public int lineSpacing;
 	public int screenOrientation;
 	
 	public ArrayList<EpubInfo> arrRecentEpub = new ArrayList<EpubInfo>(); // books that user has selected
